+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # stoker
 
+## 0.4.57
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.19
+    - @stoker-platform/web-app@0.5.48
+    - @stoker-platform/cli@0.5.42
+    - @stoker-platform/node-client@0.5.31
+    - @stoker-platform/utils@0.5.25
+    - @stoker-platform/web-client@0.5.27
+
 ## 0.4.56
 
 ### Patch Changes